backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
pydantic-settings==2.7.1
python-dateutil==2.9.0.post0
httpx==0.28.1
apscheduler==3.11.0
psycopg[binary]==3.2.4
